еще больше субд

+++ src/main/java/school2107/bot/realiser.java
@@ -56,9 +56,7 @@ public class realiser {
         addMemberOE(1l,"tst",0);
         addMemberOE(2l,"tst",0);
         addMemberOE(3l,"tst",0);
-        getMemberOE("tst",0);
         rmMemberOE(1l,"tst",0);
-        getMemberOE("tst",0);
     }
     public void createTableOE(String name){
         try {
@@ -74,6 +72,7 @@ public class realiser {
         if(oe==3){table="ege";}
         try {
             Statement s = connection.createStatement();
+            if(isMemberOE(uid,sbt,oe)){return;}
             s.execute("INSERT INTO "+table+" VALUES ("+uid+",'"+sbt+"')");
         } catch (SQLException throwables) {
             throwables.printStackTrace();
@@ -85,27 +84,49 @@ public class realiser {
         if(oe==3){table="ege";}
         try {
             Statement s = connection.createStatement();
+            if(!isMemberOE(uid,sbt,oe)){return;}
             s.execute("DELETE FROM "+table+" WHERE UID="+uid+" AND sbt='"+sbt+"'");
         } catch (SQLException throwables) {
             throwables.printStackTrace();
         }
     }
-    public ArrayList<Long> getMemberOE(String sbt,int oe){
+    public ArrayList<Long> getMembersOE(String sbt,int oe){
+        System.out.println("requested");
         String table="errhandler";
         if(oe==2){table="oge";}
         if(oe==3){table="ege";}
         ArrayList<Long> rtn = new ArrayList<>();
         try {
             Statement s = connection.createStatement();
-            ResultSet rs = s.executeQuery("SELECT SBT="+sbt+" FROM "+table);
-
+            ResultSet rs = s.executeQuery("SELECT * FROM "+table);
             while(rs.next()){
-                rtn.add(rs.getLong("uid"));
-                System.out.println(rs.getLong("uid"));
+                if(rs.getString("sbt").equalsIgnoreCase(sbt)) {
+                    System.out.println(rs.getLong("uid"));
+                    rtn.add(rs.getLong("uid"));
+                }
             }
         }catch(SQLException e){
 
         }
         return rtn;
+    }
+    public boolean isMemberOE(Long uid,String sbt,int oe){
+        System.out.println("check");
+        String table="errhandler";
+        if(oe==2){table="oge";}
+        if(oe==3){table="ege";}
+        ArrayList<Long> rtn = new ArrayList<>();
+        try {
+            Statement s = connection.createStatement();
+            ResultSet rs = s.executeQuery("SELECT * FROM "+table);
+            while(rs.next()){
+                if(rs.getString("sbt").equalsIgnoreCase(sbt)) {
+                    if(rs.getLong("uid")==uid){return true;}
+                }
+            }
+        }catch(SQLException e){
+
+        }
+        return false;
     }
 }
